example/centralized_fl/requirements.txt
torch
torchvision
scikit-learn
tqdm
matplotlib
python-dotenv
schedule
torchinfo
requests
GitPython